move time_helpers to shared_utils
