+++ _shared_utils/shared_utils/__init__.py
@@ -9,6 +9,7 @@ from . import (
     rt_dates,
     rt_utils,
     schedule_rt_utils,
+    time_helpers,
 )
 
 __all__ = [
@@ -22,4 +23,5 @@ __all__ = [
     "rt_dates",
     "rt_utils",
     "schedule_rt_utils",
+    "time_helpers",
 ]
