+++ rt_segment_speeds/segment_speed_utils/__init__.py
@@ -7,7 +7,6 @@ from . import (
     parallel_corridors,
     project_vars,
     segment_calcs,
-    time_helpers,
     time_series_utils,
     vp_transform,
 )
@@ -21,7 +20,6 @@ __all__ = [
     "parallel_corridors",
     "project_vars",
     "segment_calcs",
-    "time_helpers",
     "time_series_utils",
     "vp_transform",
 ]
